240906 김명석 || 자잘한 버그 수정
