+++ Assets/Animations/original/warrior.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8dce8c758361c0a42b4f43bc4f9f0c51
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
