+++ Assets/Animations/original/warrior/wa_atkd.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3b4809c87c39c49408c87534f8f55e42
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
