+++ Assets/Animations/original/warrior/wa_con.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2b519c3b3244b1498286ff0dbd369af
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
