+++ Assets/Animations/original/warrior/wa_running_a.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 988d1deb214788d478a2b70ed35b86cd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
