+++ Assets/Animations/original/warrior/wa_running_d.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 736b1fd409b05ae42aa0a8dab0ffaff5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
