+++ Assets/Animations/original/warrior/wa_running_s.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: baa9d244f4ccfdd41be56e42d92a8d8c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
