+++ Assets/Animations/original/warrior/wa_running_w.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c7eea6e12620fa540b8998de88046c37
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
