+++ Assets/Animations/original/warrior/wa_standing.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f58f6555eef44884687fc188332778e7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
