--- Assets/Scripts/play/button_onclick.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 75067a3b71185f446bb713a4a2714812
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
