+++ Assets/Scripts/play/engineerSkill.cs
@@ -1,11 +1,12 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
-using String = System.String;
+
 
 public class engineerSkill : MonoBehaviour, playerMove
 {
-    // Start is called before the first frame update
     Rigidbody2D rigid;
     Animator animator;  
     Stat stat;
@@ -14,6 +15,7 @@ public class engineerSkill : MonoBehaviour, playerMove
 
     private void Start() {
         Player = gameObject.GetComponent<player>();
+        stat = Player.stat;
         Player.skillMap.Add(0, "teleport");  //keyset의 인덱스 0번을 텔레포트로 지정  
         animator = gameObject.GetComponent<Animator>();
         Debug.Log("sdf");
@@ -23,9 +25,7 @@ public class engineerSkill : MonoBehaviour, playerMove
         switch(skillName){
 
             case "teleport" : 
-                Player.isInputBlocked = true;
                 StartCoroutine(player_tele());
-                Player.isInputBlocked = false;
                 break;
             default :
                 Debug.Log("존재하지 않는 스킬입니다.");
@@ -34,10 +34,14 @@ public class engineerSkill : MonoBehaviour, playerMove
         }
     }
 
-    private IEnumerator player_tele()
+    /**** 텔레포트 ****/
+    /* 시전 중 isInputBlocked를 true로 만들어 키 입력을 막음 */
+    /* layer를 변경시켜 충돌처리를 적용치않음 */
+    /* isMoving : 키 입력 여부(키를 누르고 있으면 True, 키를 뗀다면 False로 변경) */
+    private IEnumerator player_tele()  
     {   
+        float delay = 1f;
         Debug.Log("teleporting");
-        float delay = 0.01f;
         
         Vector3 moveTo;
 
@@ -58,17 +62,21 @@ public class engineerSkill : MonoBehaviour, playerMove
                 moveTo = new Vector3(0, 0, 0);
                 break;
         }
+        Player.isInputBlocked = true;
         animator.SetBool("isTele", true);
         animator.SetBool("isDirChg", false);
 
-        //delay동안  키입력을 막고 무적상태
-        while(animator.GetBool("isTele")==true){
-            yield return new WaitForSeconds(delay);
-        }
+        // //delay동안  키입력을 막고 무적상태, 나중에 애니메이션 이벤트에서 isTele를 false로 변경하는 기능 적용 후에 적용
+        // while(animator.GetBool("isTele")==true){
+        //     yield return new WaitForSeconds(delay);
+        // }
+
+        yield return new WaitForSeconds(delay);
         animator.SetBool("isDelay", false);
         animator.SetBool("isTele", false);
-        transform.position += moveTo * stat.moveSpeed * Time.deltaTime * 3;
+        transform.position += moveTo * stat.moveSpeed * Time.deltaTime * 10;
         //동작완료 0.05초 후 이동가능
         yield return new WaitForSeconds(0.2f);
+        Player.isInputBlocked = false;
     }
 }
